requirements.txt
absl-py==0.2.1
apturl==0.5.2
asn1crypto==0.24.0
astor==0.6.2
astroid==2.0
attrs==18.1.0
Automat==0.6.0
autopep8==1.3.5
backcall==0.1.0
beautifulsoup4==4.4.1
bleach==1.5.0
blinker==1.3
boto==2.48.0
boto3==1.7.50
botocore==1.10.50
Brlapi==0.6.4
bs4==0.0.1
bz2file==0.98
cffi==1.11.5
chardet==2.3.0
checkbox-support==0.22
command-not-found==0.3
constantly==15.1.0
cryptography==2.2.2
cssselect==1.0.3
cycler==0.10.0
decorator==4.3.0
defer==1.0.6
dill==0.2.7.1
docopt==0.6.2
docutils==0.14
editdistance==0.4
entrypoints==0.2.3
enum34==1.1.6
feedparser==5.1.3
future==0.16.0
gast==0.2.0
gensim==3.4.0
graphviz==0.8.4
grpcio==1.12.0
guacamole==0.9.2
gym==0.10.5
h5py==2.7.1
h5sparse==0.0.5
html5lib==0.9999999
httplib2==0.9.1
hyperlink==18.0.0
idna==2.6
incremental==17.5.0
ipykernel==4.8.2
ipython==6.4.0
ipython-genutils==0.2.0
ipywidgets==7.2.1
isort==4.3.4
jedi==0.12.0
jieba==0.39
Jinja2==2.8
jmespath==0.9.3
jsonschema==2.6.0
jupyter==1.0.0
jupyter-client==5.2.3
jupyter-console==5.2.0
jupyter-core==4.4.0
Keras==2.2.0
Keras-Applications==1.0.2
Keras-Preprocessing==1.0.1
kiwisolver==1.0.1
language-selector==0.1
lazy-object-proxy==1.3.1
louis==2.6.4
lxml==3.5.0
Markdown==2.6.11
MarkupSafe==0.23
matplotlib==2.2.2
mccabe==0.6.1
mistune==0.8.3
multiprocess==0.70.5
nbconvert==5.3.1
nbformat==4.4.0
nltk==3.3
notebook==5.5.0
numpy==1.14.3
oauthlib==1.0.3
onboard==1.2.0
padme==1.1.1
pandas==0.23.3
pandocfilters==1.4.2
parsel==1.4.0
parso==0.2.0
pexpect==4.0.1
pickleshare==0.7.4
Pillow==5.1.0
pipreqs==0.4.9
plainbox==0.25
prompt-toolkit==1.0.15
protobuf==3.5.2.post1
psutil==3.4.2
ptyprocess==0.5
pyasn1==0.1.9
pyasn1-modules==0.2.1
pycodestyle==2.4.0
pycparser==2.18
pycups==1.9.73
pycurl==7.43.0
PyDispatcher==2.0.5
pydot==1.2.4
pydot-ng==1.0.0
pygame==1.9.3
pyglet==1.3.2
Pygments==2.2.0
pygobject==3.20.0
PyJWT==1.3.0
pylint==2.0.0
pyOpenSSL==18.0.0
pypal==0.2
pyparsing==2.2.0
python-apt==1.1.0b1+ubuntu0.16.4.1
python-dateutil==2.7.3
python-debian==0.1.27
python-systemd==231
pytorch-ignite==0.1.0
pytz==2018.4
pyxdg==0.25
PyYAML==3.12
pyzmq==17.0.0
qtconsole==4.3.1
queuelib==1.5.0
reportlab==3.3.0
requests==2.9.1
s3transfer==0.1.13
scikit-learn==0.19.1
scipy==1.1.0
Scrapy==1.5.0
screen-resolution-extra==0.0.0
Send2Trash==1.5.0
service-identity==17.0.0
sessioninstaller==0.0.0
simplegeneric==0.8.1
six==1.10.0
sklearn==0.0
smart-open==1.6.0
system-service==0.3
tensorboard==1.8.0
tensorboardX==1.2
tensorflow-gpu==1.8.0
termcolor==1.1.0
terminado==0.8.1
testpath==0.3.1
torch==0.4.0
torch-vision==0.1.6.dev0
torchfile==0.1.0
torchvision==0.2.1
tornado==5.0.2
traitlets==4.3.2
Twisted==18.4.0
typed-ast==1.1.0
ubuntu-drivers-common==0.0.0
ufw==0.35
unattended-upgrades==0.1
unity-scope-calculator==0.1
unity-scope-chromiumbookmarks==0.1
unity-scope-colourlovers==0.1
unity-scope-devhelp==0.1
unity-scope-firefoxbookmarks==0.1
unity-scope-gdrive==0.7
unity-scope-manpages==0.1
unity-scope-openclipart==0.1
unity-scope-texdoc==0.1
unity-scope-tomboy==0.1
unity-scope-virtualbox==0.1
unity-scope-yelp==0.1
unity-scope-zotero==0.1
unity-tweak-tool==0.0.7
UNKNOWN==0.0.0
urllib3==1.13.1
usb-creator==0.3.0
visdom==0.1.8.4
w3lib==1.19.0
wcwidth==0.1.7
websocket-client==0.48.0
Werkzeug==0.14.1
widgetsnbextension==3.2.1
wrapt==1.10.11
xdiagnose==3.8.4.1
xkit==0.0.0
XlsxWriter==0.7.3
yarg==0.1.9
zope.interface==4.5.0
